implements csv download option to bookmarks
