+++ app/views/bookmarks/_bookmark_actions.html.erb
@@ -1,24 +1,25 @@
 <% if current_user.present? %>
-  <% if @bookmark_category.present? %>
-    <div class="shareable-url-section">
-      <div class="bookmark-action-tools">
-        <% button_text = "Generate Shared Url" %>
-        <div class="pull-left">
-          <% if @bookmark_category.access_token.present? %>
-            <% button_text = "Generate New Shared Url" %>
-            <%= link_to "Copy Url", "javascript:void(0)", class: 'copy_bookmark_url btn btn-primary', style: "color: #fff;" %>
-            <%= link_to "Delete Url", delete_share_url_bookmarks_path(bookmark_category_id: @bookmark_category.id), class: 'btn btn-danger', style: "color: #fff;", method: :delete %>
-          <% end %>
-        </div>
-        <div class="pull-right">
-          <%= link_to button_text, generate_share_url_bookmarks_path(bookmark_category_id: @bookmark_category.id), class: 'btn btn-primary', style: "color: #fff;" %>
-        </div>
+  <div class="shareable-url-section">
+    <div class="bookmark-action-tools">
+      <% button_text = "Generate Shared Url" %>
+      <div class="pull-left">
+        <% if @bookmark_category.present? && @bookmark_category.access_token.present? %>
+          <% button_text = "Generate New Shared Url" %>
+          <%= link_to "Copy Url", "javascript:void(0)", class: 'copy_bookmark_url btn btn-primary', style: "color: #fff;" %>
+          <%= link_to "Delete Url", delete_share_url_bookmarks_path(bookmark_category_id: @bookmark_category.id), class: 'btn btn-danger', style: "color: #fff;", method: :delete %>
+        <% end %>
       </div>
-      <% if @bookmark_category.access_token.present? %>
-        <div class="shareable-url">
-          <%= request.base_url + bookmarks_path(encrypted_user_id: @bookmark_category.access_token) %>
-        </div>  
-      <% end %>
+      <div class="pull-right">
+        <% if @bookmark_category.present? %>
+          <%= link_to button_text, generate_share_url_bookmarks_path(bookmark_category_id: @bookmark_category.id), class: 'btn btn-primary', style: "color: #fff;" %>
+        <% end %>
+        <%= link_to "Download Bookmark results to .CSV file", bookmarks_path(bookmark_category_id: @bookmark_category&.id, format: :csv), class: 'btn btn-primary', style: "color: #fff;" %>
+      </div>
     </div>
-  <% end %>
+    <% if @bookmark_category.present? && @bookmark_category.access_token.present? %>
+      <div class="shareable-url">
+        <%= request.base_url + bookmarks_path(encrypted_user_id: @bookmark_category.access_token) %>
+      </div>  
+    <% end %>
+  </div>
 <% end %>
